Seperate using a BaseCharacter parent class, also for controls
