+++ app/src/ts/Classes/Npc.ts
@@ -1,37 +1,26 @@
-import { IPlayer } from '../Interfaces/IPlayer.ts';
 import { ModelPrefix } from '../Enums/ModelPrefix.ts';
 import { Vector3, Quaternion } from 'three';
-import ExperienceCamera from './ExperienceCamera.ts';
 import ExperienceScene from './ExperienceScene.ts';
 import BaseCharacter from './BaseCharacter.ts';
-import PlayerControls from './PlayerControls.ts';
+import ExperienceCamera from './ExperienceCamera.ts';
+import NpcControls from './NpcControls.ts';
 
-export default class Player extends BaseCharacter implements IPlayer {
-	public isCurrent: boolean = false;
-
+export default class Npc extends BaseCharacter {
 	constructor(
 		username: string,
 		modelPrefix: ModelPrefix,
 		modelId: number,
 		experienceScene: ExperienceScene,
 		camera: ExperienceCamera,
-		isCurrent: boolean = false,
 		spawnPosition: Vector3 = new Vector3(),
 		spawnRotation: Quaternion = new Quaternion()
 	) {
 		super(username, modelPrefix, modelId, camera, experienceScene, spawnPosition, spawnRotation);
 
-		this.isCurrent = isCurrent;
-
 		// Initiate
 		this.init().then(() => {
-			if (this.isCurrent) {
-				// Setup controls after player is fully initiated
-				this.controls = new PlayerControls(this);
-
-				// Make sure controls are connected
-				this.controls.connect();
-			}
+			// Setup controls after npc is fully initiated
+			this.controls = new NpcControls(this);
 		});
 	}
 }
