+++ app/src/ts/Classes/NpcControls.ts
@@ -1,0 +1,8 @@
+import Npc from './Npc.ts';
+import BaseCharacterControls from './BaseCharacterControls.ts';
+
+export default class NpcControls extends BaseCharacterControls {
+	constructor(npc: Npc) {
+		super(npc);
+	}
+}
