+++ app/src/ts/Classes/PlayerControls.ts
@@ -1,282 +1,8 @@
-import { KeyboardKey } from '../Enums/KeyboardKey.ts';
-import { AnimationName } from '../Enums/AnimationName.ts';
-import { IPlayerControls } from '../Interfaces/IPlayerControls.ts';
-import { Quaternion, Vector3 } from 'three';
 import Player from './Player.ts';
+import BaseCharacterControls from './BaseCharacterControls.ts';
 
-export default class PlayerControls implements IPlayerControls {
-	private readonly player: Player;
-	private currentAction: AnimationName = AnimationName.IDLE;
-	private walkDirection: Vector3 = new Vector3();
-	private rotateAngle: Vector3 = new Vector3(0, 1, 0);
-	private rotateQuaternion: Quaternion = new Quaternion();
-	private isJumping: boolean = false;
-	private fadeDuration: number = 0.2;
-	private runVelocity: number = 5;
-	private walkVelocity: number = 2;
-	private emoteAnimationTimeout: ReturnType<typeof setTimeout> | null = null;
-	public emoteAnimationName: AnimationName | null = null;
-	public keysPressed: { [key in KeyboardKey]: boolean } = {} as {
-		[key in KeyboardKey]: boolean;
-	};
-	private directions: KeyboardKey[] = [
-		KeyboardKey.KeyW,
-		KeyboardKey.KeyA,
-		KeyboardKey.KeyS,
-		KeyboardKey.KeyD,
-		KeyboardKey.ArrowUp,
-		KeyboardKey.ArrowDown,
-		KeyboardKey.ArrowLeft,
-		KeyboardKey.ArrowRight
-	];
-
+export default class PlayerControls extends BaseCharacterControls {
 	constructor(player: Player) {
-		// Set properties
-		this.player = player;
-
-		if (this.player.isCurrent) {
-			// Make sure camera parent position is starts at spawn position
-			this.player.experienceScene.cameraParent.position.set(
-				this.player.spawnPosition.x,
-				this.player.spawnPosition.y,
-				this.player.spawnPosition.z
-			);
-		}
-
-		// Call animation when corresponding key is pressed
-		this.player.animationsMap.forEach((animation, key) => {
-			if (key === this.currentAction) {
-				// Play the animation
-				animation.play();
-			}
-		});
-	}
-
-	connect() {
-		// Disconnect first to make sure event listeners only exist ones
-		this.disconnect();
-
-		// Add event listeners
-		document.addEventListener('keydown', this.onKeyDown.bind(this));
-		document.addEventListener('keyup', this.onKeyUp.bind(this));
-	}
-
-	disconnect() {
-		// Remove event listeners
-		document.removeEventListener('keydown', this.onKeyDown.bind(this));
-		document.removeEventListener('keyup', this.onKeyUp.bind(this));
-	}
-
-	onKeyDown(event: KeyboardEvent) {
-		// Save pressed key
-		this.keysPressed[event.code as KeyboardKey] = true;
-	}
-
-	onKeyUp(event: KeyboardEvent) {
-		// Save pressed key
-		this.keysPressed[event.code as KeyboardKey] = false;
-	}
-
-	update(delta: number, keysPressed: { [key in KeyboardKey]: boolean }) {
-		if (!this.player.model) {
-			return;
-		}
-
-		if (this.emoteAnimationName) {
-			this.playAnimation(this.emoteAnimationName);
-
-			return;
-		}
-
-		// Check direction
-		const directionPressed = this.directions.some((key) => keysPressed[key]);
-
-		// Check if user needs to run
-		const isRunKeyPressed = keysPressed[KeyboardKey.ShiftLeft];
-
-		// Check if user needs to jump
-		const isJumpKeyPressed = keysPressed[KeyboardKey.Space];
-
-		// Determine the animation
-		let play: AnimationName | null;
-		if (directionPressed && isRunKeyPressed) {
-			play = AnimationName.RUNNING;
-
-			if (isJumpKeyPressed) {
-				play = AnimationName.RUNNING_JUMP;
-			}
-		} else if (directionPressed && !isRunKeyPressed) {
-			play = AnimationName.WALKING;
-
-			if (isJumpKeyPressed) {
-				play = AnimationName.RUNNING_JUMP;
-			}
-		} else if (isJumpKeyPressed) {
-			play = AnimationName.JUMPING;
-		} else {
-			play = AnimationName.IDLE;
-		}
-
-		// Do the animation if it's not already the current animation state
-		if (this.currentAction !== play && !this.isJumping && this.player) {
-			// Do a transition to the new animation state
-			const animationToPlay = this.player.animationsMap.get(play) ?? this.player.animationsMap.get(AnimationName.IDLE);
-			const currentAnimation = this.player.animationsMap.get(this.currentAction);
-
-			if (!animationToPlay || !currentAnimation) return;
-
-			const clip = animationToPlay.getClip();
-			const clipDuration = clip.duration * 0.8;
-
-			// User does combo's => the previously set timeout must be cleared to stop the new animation at the correct time
-			if (isJumpKeyPressed) {
-				// Clear timeout if one is already set
-				if (this.emoteAnimationTimeout) {
-					// Reset
-					clearTimeout(this.emoteAnimationTimeout);
-					this.emoteAnimationTimeout = null;
-				}
-
-				// Set jumping state
-				this.isJumping = true;
-
-				// Reset when animation is done
-				this.emoteAnimationTimeout = setTimeout(() => {
-					// Set false to disable jump or running jump animation
-					this.isJumping = false;
-				}, clipDuration * 1000);
-			}
-
-			// Fade out the current animation
-			currentAnimation.fadeOut(this.fadeDuration);
-
-			// Fade in the new animation
-			animationToPlay.reset().fadeIn(this.fadeDuration).play();
-
-			// Set the new animation state
-			this.currentAction = play;
-		}
-
-		if (
-			this.currentAction === AnimationName.RUNNING ||
-			this.currentAction === AnimationName.WALKING ||
-			this.currentAction === AnimationName.RUNNING_JUMP
-		) {
-			// Calculate camera Y angle
-			const angleYCameraAngle = this.directionOffset(keysPressed);
-
-			// Rotate player model
-			this.rotateQuaternion.setFromAxisAngle(this.rotateAngle, angleYCameraAngle);
-
-			// Smoothly rotate player model using slerp and apply delta to make it frame independent
-			this.player.model?.quaternion.slerp(this.rotateQuaternion, Math.min(3 * delta, 1));
-
-			if (this.player.isCurrent) {
-				// Calculate direction
-				this.player.camera.getWorldDirection(this.walkDirection);
-			} else {
-				const visitorWorldDirection = this.player.camera.getWorldDirection(new Vector3());
-				this.walkDirection = new Vector3(visitorWorldDirection.x, 0, visitorWorldDirection.z);
-			}
-
-			this.walkDirection.y = 0;
-			this.walkDirection.normalize();
-			this.walkDirection.applyAxisAngle(this.rotateAngle, angleYCameraAngle);
-
-			// Run/walk velocity
-			let velocity = 0;
-			switch (this.currentAction) {
-				case AnimationName.RUNNING:
-					velocity = this.runVelocity;
-					break;
-				case AnimationName.RUNNING_JUMP:
-					velocity = this.runVelocity;
-					break;
-				case AnimationName.WALKING:
-					velocity = this.walkVelocity;
-					break;
-			}
-
-			// Move player & camera
-			const moveX = this.walkDirection.x * velocity * delta;
-			const moveZ = this.walkDirection.z * velocity * delta;
-
-			// Update player position
-			this.player.model.position.x += moveX;
-			this.player.model.position.z += moveZ;
-
-			if (this.player.isCurrent) {
-				// Sync camera with current player player
-				this.player.experienceScene.cameraParent.position.x = this.player.model.position.x;
-				this.player.experienceScene.cameraParent.position.z = this.player.model.position.z;
-			}
-		}
-	}
-
-	playAnimation(animationName: AnimationName) {
-		if (this.currentAction !== animationName) {
-			const animationToPlay = this.player.animationsMap.get(animationName);
-			const currentAnimation = this.player.animationsMap.get(this.currentAction);
-
-			if (!animationToPlay || !currentAnimation) {
-				return;
-			}
-
-			// Clear the previous timeout if it's set
-			if (this.emoteAnimationTimeout) {
-				clearTimeout(this.emoteAnimationTimeout);
-				this.emoteAnimationTimeout = null;
-			}
-
-			// Get the duration of the animation (you may need to adjust this depending on your system)
-			const clip = animationToPlay.getClip();
-			const clipDuration = clip.duration * 0.8; // Optional: Adjust duration if needed
-
-			// Set a timeout to reset the emote animation name after the animation duration
-			this.emoteAnimationTimeout = setTimeout(() => {
-				this.emoteAnimationName = null; // Reset the emote animation
-			}, clipDuration * 1000); // Multiply by 1000 to convert to milliseconds
-
-			// Fade out the current animation
-			currentAnimation.fadeOut(this.fadeDuration);
-
-			// Reset and fade in the new animation
-			animationToPlay.reset().fadeIn(this.fadeDuration).play();
-			this.currentAction = animationName;
-		}
-	}
-
-	directionOffset(keysPressed: { [key in KeyboardKey]: boolean }) {
-		// z
-		let directionOffset = 0;
-
-		if (keysPressed[KeyboardKey.KeyW] || keysPressed[KeyboardKey.ArrowUp]) {
-			if (keysPressed[KeyboardKey.KeyA] || keysPressed[KeyboardKey.ArrowLeft]) {
-				// z+q
-				directionOffset = Math.PI / 4;
-			} else if (keysPressed[KeyboardKey.KeyD] || keysPressed[KeyboardKey.ArrowRight]) {
-				// z+d
-				directionOffset = -Math.PI / 4;
-			}
-		} else if (keysPressed[KeyboardKey.KeyS] || keysPressed[KeyboardKey.ArrowDown]) {
-			if (keysPressed[KeyboardKey.KeyA] || keysPressed[KeyboardKey.ArrowLeft]) {
-				// s+q
-				directionOffset = Math.PI / 4 + Math.PI / 2;
-			} else if (keysPressed[KeyboardKey.KeyD] || keysPressed[KeyboardKey.ArrowRight]) {
-				// s+d
-				directionOffset = -Math.PI / 4 - Math.PI / 2;
-			} else {
-				// s
-				directionOffset = Math.PI;
-			}
-		} else if (keysPressed[KeyboardKey.KeyA] || keysPressed[KeyboardKey.ArrowLeft]) {
-			// q
-			directionOffset = Math.PI / 2;
-		} else if (keysPressed[KeyboardKey.KeyD] || keysPressed[KeyboardKey.ArrowRight]) {
-			// d
-			directionOffset = -Math.PI / 2;
-		}
-
-		return directionOffset;
+		super(player);
 	}
 }
