+++ app/src/ts/Interfaces/INpc.ts
@@ -1,0 +1,5 @@
+import ExperienceCamera from '../Classes/ExperienceCamera.ts';
+
+export interface INpc {
+	camera: ExperienceCamera;
+}
