+++ app/src/ts/Interfaces/IPlayer.ts
@@ -1,8 +1,3 @@
-import ExperienceCamera from '../Classes/ExperienceCamera.ts';
-import PlayerControls from '../Classes/PlayerControls.ts';
-
 export interface IPlayer {
-	camera: ExperienceCamera;
 	isCurrent: boolean;
-	controls: PlayerControls | null;
 }
